+++ firebase-vscode/src/logger-wrapper.ts
@@ -1,5 +1,6 @@
 import * as path from "path";
 import * as vscode from "vscode";
+import * as fs from "fs";
 import { transports, format } from "winston";
 import Transport from "winston-transport";
 import stripAnsi from "strip-ansi";
@@ -45,7 +46,11 @@ export function logSetup() {
     // Re-implement file logger call from ../../src/bin/firebase.ts to not bring
     // in the entire firebase.ts file
     const rootFolders = getRootFolders();
-    const filePath = path.join(rootFolders[0], ".firebase", "logs", "vsce-debug.log");
+    // Default to a central path, but write files to a local path if we're in a Firebase directory.
+    let filePath = "~/.cache/firebase/vscode/vsce-debug.log";
+    if (fs.existsSync(path.join(rootFolders[0], "firebase.json")) { 
+      filePath = path.join(rootFolders[0], ".firebase", "logs", "vsce-debug.log");
+    }
     pluginLogger.info("Logging to path", filePath);
     cliLogger.add(
       new transports.File({